+++ feature/doglist/src/main/java/com/sidgowda/pawcalc/doglist/navigation/DogListNavigation.kt
@@ -14,7 +14,7 @@ fun NavController.navigateToDogListScreen(navOptions: NavOptions? = null) {
 fun NavGraphBuilder.dogListScreenDestination(
     onNavigateToOnboarding: () -> Unit,
     onNewDog: () -> Unit,
-    onEditDog: (Long) -> Unit,
+    onEditDog: (Int) -> Unit,
     onDogDetails: () -> Unit
 ) {
     composable(route = DOG_LIST_SCREEN_ROUTE) { backStackEntry ->
